# Optimize bootstrapping code by lazily loading classes not needed on every request