--- a/.github/workflows/build-test-lint.yml
+++ b/.github/workflows/build-test-lint.yml
@@ -45,5 +45,4 @@ jobs:
         run: npm run lint
 
       - name: Run benchmark
-        run: php tests/benchmarks/plugin.php
-
+        run: php tests/benchmarks/run-benchmark.php