--- a/koko-analytics.php
+++ b/koko-analytics.php
@@ -47,16 +47,22 @@ if (PHP_VERSION_ID < 70400 || ! \defined('ABSPATH')) {
     return;
 }
 
+$times = [];
+$times[] = ['start', microtime(true)];
+
 // Maybe run any pending database migrations
 require __DIR__ . '/src/class-migrations.php';
 $migrations = new Migrations('koko_analytics_version', KOKO_ANALYTICS_VERSION, KOKO_ANALYTICS_PLUGIN_DIR . '/migrations/');
 add_action('init', [$migrations, 'maybe_run'], 10, 0);
+$times[] = ['Migrations', microtime(true)];
 
 require __DIR__ . '/src/class-aggregator.php';
 new Aggregator();
+$times[] = ['Aggregator', microtime(true)];
 
 require __DIR__ . '/src/class-plugin.php';
 new Plugin();
+$times[] = ['Plugin', microtime(true)];
 
 if (\defined('DOING_AJAX') && DOING_AJAX) {
     // ajax only
@@ -70,33 +76,65 @@ if (\defined('DOING_AJAX') && DOING_AJAX) {
     require __DIR__ . '/src/class-script-loader.php';
     new Script_Loader();
     add_action('admin_bar_menu', 'KokoAnalytics\admin_bar_menu', 40, 1);
+    $times[] = ['Script_Loader', microtime(true)];
 }
 
 require __DIR__ . '/src/class-query-loop-block.php';
 new QueryLoopBlock();
+$times[] = ['QueryLoopBlock', microtime(true)];
 
-require __DIR__ . '/src/class-dashboard.php';
-new Dashboard();
-
-require __DIR__ . '/src/class-rest.php';
-new Rest();
-
-require __DIR__ . '/src/class-shortcode-most-viewed-posts.php';
-new Shortcode_Most_Viewed_Posts();
-
-require __DIR__ . '/src/class-shortcode-site-counter.php';
-new ShortCode_Site_Counter();
+// init REST API endpoint
+add_action('rest_api_init', [Rest::class, 'register_routes'], 10, 0);
+$times[] = ['Rest', microtime(true)];
 
 require __DIR__ . '/src/class-pruner.php';
 new Pruner();
+$times[] = ['Pruner', microtime(true)];
 
 if (\class_exists('WP_CLI')) {
     \WP_CLI::add_command('koko-analytics', 'KokoAnalytics\Command');
+    $times[] = ['Command', microtime(true)];
 }
 
+
+
+// register shortcodes
+add_shortcode('koko_analytics_most_viewed_posts', [Shortcode_Most_Viewed_Posts::class, 'content']);
+$times[] = ['Shortcode_Most_Viewed_Posts', microtime(true)];
+
+add_shortcode('koko_analytics_counter', [Shortcode_Site_Counter::class, 'content']);
+$times[] = ['ShortCode_Site_Counter', microtime(true)];
+
+// maybe show standalone dashboard
+add_action('wp', function () {
+    if (!isset($_GET['koko-analytics-dashboard'])) {
+        return;
+    }
+
+    $settings = get_settings();
+    if (!$settings['is_dashboard_public'] && !current_user_can('view_koko_analytics')) {
+        return;
+    }
+
+    (new Dashboard())->show_standalone_dashboard_page();
+});
+$times[] = ['Dashboard', microtime(true)];
+
+// register most viewed posts widget
 add_action('widgets_init', function () {
     require KOKO_ANALYTICS_PLUGIN_DIR . '/src/class-widget-most-viewed-posts.php';
     register_widget(Widget_Most_Viewed_Posts::class);
 });
 
 add_action('koko_analytics_test_custom_endpoint', 'KokoAnalytics\test_custom_endpoint');
+
+// $times = array_map(function ($t) {
+//     $t[1] = round($t[1] * 1000, 4);
+//     return $t;
+// }, $times);
+
+// for ($i = count($times) - 1; $i > 0; $i--) {
+//     $times[$i][1] = $times[$i][1] - $times[$i-1][1];
+// }
+
+// dump($times);