--- a/src/class-pruner.php
+++ b/src/class-pruner.php
@@ -10,26 +10,19 @@ namespace KokoAnalytics;
 
 class Pruner
 {
-    public function __construct()
-    {
-        add_action('koko_analytics_prune_data', [$this, 'run'], 10, 0);
-        register_activation_hook(KOKO_ANALYTICS_PLUGIN_FILE, [$this, 'setup_scheduled_event']);
-        register_deactivation_hook(KOKO_ANALYTICS_PLUGIN_FILE, [$this, 'clear_scheduled_event']);
-    }
-
-    public function setup_scheduled_event(): void
+    public static function setup_scheduled_event(): void
     {
         if (! wp_next_scheduled('koko_analytics_prune_data')) {
             wp_schedule_event(time() + DAY_IN_SECONDS, 'daily', 'koko_analytics_prune_data');
         }
     }
 
-    public function clear_scheduled_event(): void
+    public static function clear_scheduled_event(): void
     {
         wp_clear_scheduled_hook('koko_analytics_prune_data');
     }
 
-    public function run()
+    public static function run()
     {
         /** @var \wpdb $wpdb */
         global $wpdb;