Average across 50 runs with and without OpCache:

**Baseline**

| Name  | Time (ms) |
| ------------- | ------------- |
| Without opcache  | 4.091 ms  |
| with opcache | 0.191 ms  |

**This branch**

| Name  | Time (ms) |
| ------------- | ------------- |
| Without opcache  | 1.939 ms  |
| with opcache | 0.143 ms  |